added crc32 implementation (thanks Gary Brown)

+++ include/libnet/crc.h
@@ -4,6 +4,9 @@
 // Copyright (C) 2017 Return Infinity -- see LICENSE
 // =========================================================
 
+#include <stdint.h>
+#include <stdlib.h>
+
 #ifndef LIBNET_CRC_H
 #define LIBNET_CRC_H
 
@@ -13,12 +16,12 @@ extern "C"
 #endif
 
 /// @brief Calculates the 16-bit CRC for a block of memory.
-unsigned short int libnet_crc16(const void *buf,
-                                unsigned long long int buf_size);
+uint16_t libnet_crc16(const void *buf,
+                      size_t buf_size);
 
 /// @brief Calculates the 32-bit CRC for a block of memory.
-unsigned long int libnet_crc32(const void *buf,
-                               unsigned long long int buf_size);
+uint32_t libnet_crc32(const void *buf,
+                      size_t buf_size);
 
 #ifdef __cplusplus
 } // extern "C"
